+++ covalent_ecs_plugin/assets/infra/main.tf
@@ -23,9 +23,18 @@ resource "aws_s3_bucket" "bucket" {
   force_destroy = true
 }
 
-resource "aws_s3_bucket_acl" "bucket_acl" {
+resource "aws_s3_bucket_ownership_controls" "ownership_controls" {
   bucket = aws_s3_bucket.bucket.id
-  acl    = "private"
+
+  rule {
+    object_ownership = "BucketOwnerPreferred"
+  }
+}
+
+resource "aws_s3_bucket_acl" "bucket_acl" {
+  depends_on = [aws_s3_bucket_ownership_controls.ownership_controls]
+  bucket     = aws_s3_bucket.bucket.id
+  acl        = "private"
 }
 
 resource "aws_ecr_repository" "ecr_repository" {