+++ covalent_ecs_plugin/assets/infra/variables.tf
@@ -14,7 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-variable "name" {
+variable "prefix" {
   default     = "covalent-ecs-ft"
   description = "Name used to prefix AWS resources"
 }
@@ -37,4 +37,39 @@ variable "subnet_id" {
 variable "vpc_cidr" {
   default     = "10.0.0.0/24"
   description = "VPC CIDR range"
+}
+
+variable "cache_dir" {
+  type        = string
+  default     = "/tmp/covalent"
+  description = "Path on local machine where temporary files are generated"
+}
+
+variable "poll_freq" {
+  type        = number
+  default     = 5
+  description = "Frequency with which to poll AWS batch for the result object"
+}
+
+variable "vcpus" {
+  type        = number
+  default     = 2
+  description = "Number of vcpus a batch job will consume by default"
+}
+
+variable "memory" {
+  type        = number
+  default     = 2
+  description = "Memory in GB for the batch job"
+}
+
+variable "credentials" {
+  type        = string
+  default     = ""
+  description = "Path to the AWS shared configuration file"
+}
+
+variable "profile" {
+  type        = string
+  description = "AWS profile used during execution"
 }